Refactor seeding and server to separate concerns
