+++ db/db.js
@@ -1,12 +1,6 @@
 const Sequelize = require('sequelize');
-const mysql = require('mysql');
 
-const connection = mysql.createConnection({
-  host: 'localhost',
-  user: 'root',
-  password: 'password',
-});
-
+// change user: 'root' and password: 'password' with your credentials
 const sequelize = new Sequelize('reservations', 'root', 'password', {
   host: 'localhost',
   dialect: 'mysql',
@@ -17,12 +11,52 @@ sequelize
   .authenticate()
   .then(() => {
     console.log('db connection has been established successfully.');
-    require('./seed.js');
   })
   .catch((err) => {
     console.error('Unable to connect to the database:', err);
   });
 
-connection.end();
+const Availability = sequelize.define('restaurant',
+  {
+    name: {
+      type: Sequelize.STRING,
+    },
+    '6PM': {
+      type: Sequelize.INTEGER,
+    },
+    '6:15PM': {
+      type: Sequelize.INTEGER,
+    },
+    '6:30PM': {
+      type: Sequelize.INTEGER,
+    },
+    '6:45PM': {
+      type: Sequelize.INTEGER,
+    },
+    '7PM': {
+      type: Sequelize.INTEGER,
+    },
+    '7:15PM': {
+      type: Sequelize.INTEGER,
+    },
+    '7:30PM': {
+      type: Sequelize.INTEGER,
+    },
+    '7:45PM': {
+      type: Sequelize.INTEGER,
+    },
+    '8PM': {
+      type: Sequelize.INTEGER,
+    },
+    '8:15PM': {
+      type: Sequelize.INTEGER,
+    },
+    '8:30PM': {
+      type: Sequelize.INTEGER,
+    },
+  },
+  {
+    timestamps: false,
+  });
 
-module.exports = sequelize;
+module.exports = Availability;
