+++ db/seed.js
@@ -1,52 +1,7 @@
-const Sequelize = require('sequelize');
 const faker = require('faker');
 
-const sequelize = require('./db.js');
+const Availability = require('./db.js');
 
-const Availability = sequelize.define('restaurant',
-  {
-    name: {
-      type: Sequelize.STRING,
-    },
-    '6PM': {
-      type: Sequelize.INTEGER,
-    },
-    '6:15PM': {
-      type: Sequelize.INTEGER,
-    },
-    '6:30PM': {
-      type: Sequelize.INTEGER,
-    },
-    '6:45PM': {
-      type: Sequelize.INTEGER,
-    },
-    '7PM': {
-      type: Sequelize.INTEGER,
-    },
-    '7:15PM': {
-      type: Sequelize.INTEGER,
-    },
-    '7:30PM': {
-      type: Sequelize.INTEGER,
-    },
-    '7:45PM': {
-      type: Sequelize.INTEGER,
-    },
-    '8PM': {
-      type: Sequelize.INTEGER,
-    },
-    '8:15PM': {
-      type: Sequelize.INTEGER,
-    },
-    '8:30PM': {
-      type: Sequelize.INTEGER,
-    },
-  },
-  {
-    timestamps: false,
-  });
-
-// Generate random number for total seats available per restaurant per time slot
 const seats = () => faker.random.number({
   min: 2,
   max: 20,
@@ -85,6 +40,7 @@ Availability.sync({ force: true })
       });
     }
     console.log('seeded 100 restaurants');
+  })
+  .catch((err) => {
+    console.log('unable to seed db: ', err);
   });
-
-module.exports = Availability;
