+++ server/server.js
@@ -1,8 +1,7 @@
 const express = require('express');
 const bodyParser = require('body-parser');
 
-const sequelize = require('../db/db.js');
-const Availability = require('../db/seed.js');
+const Availability = require('../db/db.js');
 
 const app = express();
 const port = 3020;
@@ -15,7 +14,6 @@ app.use(express.static('public'));
 app.get('/reservations', (req, res) => {
   Availability.findOne({ where: { name: 'Kinjo' } })
     .then((main) => {
-      console.log(main);
       res.status(200).send(main);
     })
     .catch((err) => {
